requirements.txt
openai>=1.0.0
speechrecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.3.0
playsound>=1.3.0
wikipedia>=1.4.0
langgraph>=0.0.50
langchain>=0.2.0
langchain_community
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
